Moved the mdx post to an actual page
