+++ next.config.js
@@ -1,5 +1,6 @@
-const withMDX = require('@zeit/next-mdx')()
+const withMDX = require('@next/mdx')()
 
 module.exports = withMDX({
-  target: 'serverless'
+  target: 'serverless',
+  pageExtensions: ['jsx', 'js', 'ts', 'tsx', 'mdx']
 })
