+++ package.json
@@ -22,8 +22,9 @@
     ]
   },
   "dependencies": {
-    "@mdx-js/mdx": "^0.16.6",
-    "@zeit/next-mdx": "^1.2.0",
+    "@mdx-js/loader": "^1.3.1",
+    "@mdx-js/react": "^1.3.1",
+    "@next/mdx": "^9.0.4",
     "babel-plugin-inline-react-svg": "^1.0.1",
     "escape-html": "^1.0.3",
     "isomorphic-unfetch": "^3.0.0",
